+++ znjf-external-service/src/main/java/com/znjf33/external/api/provider/domain/SupplyChainLmjParamDO.java
@@ -67,7 +67,7 @@ public class SupplyChainLmjParamDO {
     /**
      * 支用申请金额
      */
-    private Float drawAmount;
+    private Double drawAmount;
     /**
      * 分期数，随借随还业务为0
      */
@@ -197,11 +197,11 @@ public class SupplyChainLmjParamDO {
         this.memberNo = memberNo;
     }
 
-    public Float getDrawAmount() {
+    public Double getDrawAmount() {
         return drawAmount;
     }
 
-    public void setDrawAmount(Float drawAmount) {
+    public void setDrawAmount(Double drawAmount) {
         this.drawAmount = drawAmount;
     }
